src/app/requirements.txt
fastapi>=0.115.0
uvicorn>=0.24.0
databricks-sdk>=0.38.0
databricks-sql-connector>=3.0.0
databricks-vectorsearch>=0.40
openai>=1.0.0
pydantic>=2.0.0
